pyproject.toml
[project]
name = "ygo-pipelines"
version = "0.1.0"
description = "Data ingestion pipelines for Yu-Gi-Oh! card data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.22",
    "pydantic>=2.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "supabase>=2.27.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
ygo-pipelines = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
